// src/lib/types.tsx
// Generated by https://quicktype.io

export interface CloudinaryPhoto {
  id: number;
  public_id: string;
  version: string;
  width: number;
  height: number;
  format: string;
  resource_type: string;
  transformation: Transformation;
  path: string;
  preview_url: string;
}

export interface Transformation {
  width: string;
  height: string;
  x: string;
  y: string;
  crop: string;
  angle: number;
}

export type Listing = {
  id: string;
  title: string;
  published_at: string;
  photos: Array<{ _links: { thumbnail: { href: string } } }>;
};